Split directives definition and parser
