+++ src/diagnostics/unknown-theme.ts
@@ -1,5 +1,5 @@
 import { Diagnostic, DiagnosticSeverity, Range, TextDocument } from 'vscode'
-import { DirectiveParser } from '../directive-parser'
+import { DirectiveParser } from '../directives/parser'
 import themes from '../themes'
 
 interface ParsedThemeValue {
